requirements.txt
ytmusicapi
google-api-python-client
python-dotenv